requirements.txt
anyio==4.9.0
apsw==3.49.1.0
apswutils==0.0.2
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
exceptiongroup==1.2.2
fastcore==1.8.0
fastlite==0.1.2
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.2
openai-whisper==20240930
packaging==24.2
pycryptodome==3.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.12.4
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
youtube-transcript-api==1.1.0
yt-dlp==2025.2.19
